--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -4,6 +4,7 @@ on: [push, pull_request, workflow_dispatch]
 
 env:
   FORCE_COLOR: 1
+  PIP_DISABLE_PIP_VERSION_CHECK: 1
 
 permissions:
   contents: read
@@ -17,4 +18,12 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: "3.x"
+          cache: pip
       - uses: pre-commit/action@v3.0.1
+
+      - name: Install dependencies
+        run: |
+          python3 -m pip install -U tox
+
+      - name: Mypy
+        run: tox -e mypy