--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -21,9 +21,17 @@ jobs:
           cache: pip
       - uses: pre-commit/action@v3.0.1
 
+  mypy:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.x"
+          cache: pip
       - name: Install dependencies
         run: |
           python3 -m pip install -U tox
-
       - name: Mypy
         run: tox -e mypy