--- a/pyproject.toml
+++ b/pyproject.toml
@@ -55,6 +55,9 @@ scripts.pep = "pepotron.cli:main"
 [tool.hatch]
 version.source = "vcs"
 
+[tool.hatch.build.hooks.vcs]
+version-file = "src/pepotron/_version.py"
+
 [tool.hatch.version.raw-options]
 local_scheme = "no-local-version"