--- a/tox.ini
+++ b/tox.ini
@@ -41,3 +41,15 @@ pass_env =
     PRE_COMMIT_COLOR
 commands =
     pre-commit run --all-files --show-diff-on-failure
+
+[testenv:mypy]
+deps =
+    mypy==1.11.2
+    platformdirs
+    pytest
+    python-slugify
+    rapidfuzz
+    types-freezegun
+    urllib3
+commands =
+    mypy --strict --pretty --show-error-codes . {posargs}