Saves about 15 ms, about twice as fast.

```console
❯ python --version
Python 3.13.0rc1
❯ python -X importtime -c "import pepotron" 2> import.log && tuna import.log
```


Before:

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/cdd4b5fc-5893-4689-9bcf-3663ef8df5c9">

After:

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/9ac6696b-4f59-4c95-8d27-bfc901a54418">

Also move mypy from pre-commit to tox, because it's easier to run it against an installed project, so `_version` has been built.


```console
❯ hyperfine --warmup 32 \
--prepare "git checkout rm-importlib.metadata" 'python3 -c "import pepotron"' \
--prepare "git checkout main"                  'python3 -c "import    pepotron"'
Benchmark 1: python3 -c "import pepotron"
  Time (mean ± σ):      25.8 ms ±   1.1 ms    [User: 20.8 ms, System: 4.4 ms]
  Range (min … max):    24.6 ms …  30.3 ms    86 runs

Benchmark 2: python3 -c "import    pepotron"
  Time (mean ± σ):      41.6 ms ±   1.4 ms    [User: 34.2 ms, System: 6.8 ms]
  Range (min … max):    40.8 ms …  50.5 ms    47 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (50.5 ms). This could be caused by (filesystem) caches that were not filled until after the first run. You are already using both the '--warmup' option as well as the '--prepare' option. Consider re-running the benchmark on a quiet system. Maybe it was a random outlier. Alternatively, consider increasing the warmup count.

Summary
  python3 -c "import pepotron" ran
    1.61 ± 0.09 times faster than python3 -c "import    pepotron"
```